requirements.txt
Django>=5.2.6
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
django-crispy-forms>=2.4
crispy-bootstrap5>=1.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
python-decouple>=3.8
celery>=5.3.0
redis>=5.0.0
django-filter>=23.5
psycopg2-binary>=2.9.9

